+++ sezzle-sdk/src/main/kotlin/com/sezzle/sdk/checkout/CheckoutState.kt
@@ -1,5 +1,6 @@
 package com.sezzle.sdk.checkout
 
+import android.net.Uri
 import com.sezzle.sdk.SezzleCheckoutListener
 
 /**
@@ -12,11 +13,15 @@ import com.sezzle.sdk.SezzleCheckoutListener
 internal object CheckoutState {
     var listener: SezzleCheckoutListener? = null
     var orderUUID: String? = null
+    var completeUrl: Uri? = null
+    var cancelUrl: Uri? = null
     var launchingActivityClassName: String? = null
 
     fun clear() {
         listener = null
         orderUUID = null
+        completeUrl = null
+        cancelUrl = null
         launchingActivityClassName = null
     }
 }